Annotazioni Payment

+++ Filiera_Francoletti_Belardinelli_Raiola/src/main/java/com/example/filiera_francoletti_belardinelli_raiola/Controller/HandlerCarrello.java
@@ -25,7 +25,7 @@ public class HandlerCarrello {
         this.products = products;
     }
 
-    public void addProduct(int id) {
+    public void addProduct(Long id) {
         Piattaforma pf = Piattaforma.getPlatform();
         Prodotto found = pf.getProductByID(id);
         if (found != null) {
