+++ Filiera_Francoletti_Belardinelli_Raiola/src/main/java/com/example/filiera_francoletti_belardinelli_raiola/Model/Users/Acquirente.java
@@ -32,7 +32,7 @@ public class Acquirente extends UtenteGenerico {
     }
 
     // Metodi di business delegati al buyerHandler
-    public void addProduct(int id) {
+    public void addProduct(Long id) {
         buyerHandler.addProduct(id);
     }
